refactor(nav): adicionar responsividade
