--- src/components/Header/Header.stories.tsx
@@ -1,9 +1,0 @@
-import { Meta } from '@storybook/react'
-import { Header } from './Header'
-
-export default {
-  title: 'Components/Header',
-  component: Header
-} as Meta
-
-export const Normal = () => <Header />
