+++ src/components/Nav/Nav.stories.tsx
@@ -1,0 +1,9 @@
+import { Meta } from '@storybook/react'
+import { Nav } from './Nav'
+
+export default {
+  title: 'Components/Nav',
+  component: Nav
+} as Meta
+
+export const Normal = () => <Nav />
