--- src/components/NavMenu/NavMenu.stories.tsx
@@ -1,9 +1,0 @@
-import { Meta } from '@storybook/react'
-import { NavMenu } from './NavMenu'
-
-export default {
-  title: 'Components/NavMenu',
-  component: NavMenu
-} as Meta
-
-export const Normal = () => <NavMenu />
