+++ src/design-system/Typography/H2.tsx
@@ -10,7 +10,7 @@ export const H2 = styled.h2<IFont>`
     ${{ ...props } as any}
   `}
 
-  @media screen and (min-width: 800px) {
+  @media screen and (min-width: 768px) {
     font-size: 25px;
   }
 `
