+++ src/design-system/Typography/H4.tsx
@@ -10,7 +10,7 @@ export const H4 = styled.h4<IFont>`
     ${{ ...props } as any}
   `}
 
-  @media screen and (min-width: 800px) {
+  @media screen and (min-width: 768px) {
     font-size: 21px;
   }
 `
