+++ src/design-system/Typography/P.tsx
@@ -10,7 +10,7 @@ export const P = styled.p<IFont>`
     ${{ ...props } as any}
   `}
 
-  @media screen and (min-width: 800px) {
+  @media screen and (min-width: 768px) {
     font-size: 18px;
   }
 `
