+++ src/layouts/DefaultLayout/index.tsx
@@ -2,7 +2,17 @@ import styled from 'styled-components'
 
 export const DefaultLayout = styled.div`
   min-height: 100vh;
-  max-width: 1280px;
   padding: 10px;
   overflow: hidden;
+
+  @media screen and (min-width: 768px) {
+    max-width: 1280px;
+    display: flex;
+    justify-content: center;
+    margin: 0 auto;
+    main {
+      margin-left: 100px;
+      max-width: 700px;
+    }
+  }
 `
